Final_Changes

+++ app/Http/Controllers/UserController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Hash;
 use App\Don;
 use App\User;
 use App\Adresse;
@@ -35,21 +36,6 @@ class UserController extends Controller
 
 
     public function updateUser(Request $request, User $user){
-        // $role = 0;
-        // if($request->role){
-        //     if($request->role == 'Teleoperateur'){
-        //         $role = 0;
-        //     }
-        //     if($request->role == 'UTSS'){
-        //         $role = 1;
-        //     }
-        //     if($request->role == 'CRRC'){
-        //         $role = 2;
-        //     }
-        //     if($request->role == 'Admin_Ministere'){
-        //         $role = 3;
-        //     }
-        // }
         try{
             if($user->id){
                 $user->update([ 'name' => $request->name,
@@ -71,13 +57,15 @@ class UserController extends Controller
         }catch(Exception $e){
             return response()->json($e->getMessage(), $e->getHttpStatus());
         }
-
     }
 
     public function deleteUser(Request $request, User $user){
         if($user->id){
-            $user->delete();
-            return response()->json('User supprime!', 200);  
+            if($user->id != 1){
+                $user->delete();
+                return response()->json('User supprime!', 200);  
+            }
+            return response()->json('Vous ne pouvez pas supprimer cet utilisateur!', 400);  
         }
         return response()->json('User non trouve!', 400);  
     }
@@ -88,5 +76,20 @@ class UserController extends Controller
             return response()->json($dons, 200);  
         }
         return response()->json('User non trouve!', 400);  
+    }
+
+    public function updatePassword(Request $request, User $user){
+        try{
+            if($user->id){
+                $user->update([ 
+                'password' => Hash::make($request->password)
+                        ]);
+                return response()->json('Mot de passe modifié!', 200);     
+            }else{
+                return response()->json('User non trouve!', 400); 
+            }
+        }catch(Exception $e){
+            return response()->json($e->getMessage(), $e->getHttpStatus());
+        }
     }
 }
